windsurf Tanisha/shapex_ar_app/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
trimesh==4.4.9
numpy==2.0.2
pydantic==2.9.2
Pillow==10.4.0
scipy==1.13.1
